Move vector sets structs to separate module.
